+++ src/components/ColorSwatch/stories/Color.story.tsx
@@ -1,6 +1,6 @@
-import { StoryObj } from "@storybook/preact"
+import { StoryObj } from '@storybook/preact'
 
-import { ColorSwatch } from "../ColorSwatch"
+import { ColorSwatch } from '../ColorSwatch'
 
 type Story = StoryObj<typeof ColorSwatch>
 
@@ -8,22 +8,22 @@ export const ColorStory: Story = {
   parameters: {
     controls: { disable: true },
     viewport: {
-      defaultViewport: "large",
+      defaultViewport: 'large',
     },
     docs: {
       source: {
-        language: "tsx",
+        language: 'tsx',
         code: `
-<ColorSwatch color={{ r: 255, g: 0, b: 0, a: 1 }}/>
+<ColorSwatch fill={{ r: 255, g: 0, b: 0, a: 1 }}/>
 `,
       },
     },
   },
   render: () => (
     <div className="sb-row sb-width-full sb-gap-16">
-      <ColorSwatch color={{ r: 255, g: 0, b: 0, a: 1 }} />
-      <ColorSwatch color={{ r: 255, g: 0, b: 0, a: 0.5 }} />
-      <ColorSwatch color={{ r: 255, g: 0, b: 0, a: 0 }} />
+      <ColorSwatch fill={{ r: 255, g: 0, b: 0, a: 1 }} />
+      <ColorSwatch fill={{ r: 255, g: 0, b: 0, a: 0.5 }} />
+      <ColorSwatch fill={{ r: 255, g: 0, b: 0, a: 0 }} />
     </div>
   ),
 }