+++ src/components/ScrollContainer/stories/Popover.story.tsx
@@ -1,0 +1,81 @@
+import { StoryObj } from '@storybook/preact'
+import { useRef } from 'preact/hooks'
+
+import { Text, ScrollContext, Stack, Popover, Button } from '../../../index'
+
+import { ScrollContainer } from '../ScrollContainer'
+
+type Story = StoryObj<typeof ScrollContainer>
+
+const initialContent = `Sed do eiusmod tempor
+              incididunt ut labore et dolore magna aliqua. Ut enim ad minim
+              veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip
+              ex ea commodo consequat. Lorem ipsum dolor sit amet, consectetur
+              adipiscing elit. Sed do eiusmod tempor incididunt ut labore et
+              dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
+              exercitation ullamco laboris nisi ut aliquip ex ea commodo
+              consequat. Lorem ipsum dolor sit amet, consectetur adipiscing
+              elit. Sed do eiusmod tempor incididunt ut labore et dolore magna
+              aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco
+              laboris nisi ut aliquip ex ea commodo consequat. Lorem ipsum dolor
+              sit amet, consectetur adipiscing elit. Sed do eiusmod tempor
+              incididunt ut labore et dolore magna aliqua. Ut enim ad minim
+              veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip
+              ex ea commodo consequat. Lorem ipsum dolor sit amet, consectetur
+              adipiscing elit. Sed do eiusmod tempor incididunt ut labore et
+              dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
+              exercitation ullamco laboris nisi ut aliquip ex ea commodo
+              consequat. Lorem ipsum dolor sit amet, consectetur adipiscing
+              elit. Sed do eiusmod tempor incididunt ut labore et dolore magna
+              aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco
+              laboris nisi ut aliquip ex ea commodo consequat. Lorem ipsum dolor
+              sit amet, consectetur adipiscing elit. Sed do eiusmod tempor
+              incididunt ut labore et dolore magna aliqua. Ut enim ad minim
+              veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip
+              ex ea commodo consequat. Lorem ipsum dolor sit amet, consectetur
+              adipiscing elit. Sed do eiusmod tempor incididunt ut labore et
+              dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
+              exercitation ullamco laboris nisi ut aliquip ex ea commodo
+              consequat. Lorem ipsum dolor sit amet, consectetur adipiscing
+              elit. Sed do eiusmod tempor incididunt ut labore et dolore magna
+              aliqua.`
+
+export const PopoverStory: Story = {
+  parameters: {
+    controls: { disable: true },
+    viewport: {
+      defaultViewport: 'large',
+    },
+    docs: {
+      source: {
+        language: 'tsx',
+        code: `
+<ScrollContext>
+  <ScrollContainer>{children}</ScrollContainer>
+</ScrollContext>
+`,
+      },
+    },
+  },
+  render: () => {
+    const triggerRef = useRef<HTMLButtonElement | null>(null)
+
+    return (
+      <div className="sb-column sb-width-full sb-height-300">
+        <Popover triggerRef={triggerRef} width={300} height={300} popoverHeaderProps={{ children: 'Popover Header' }}>
+          <ScrollContext>
+            <ScrollContainer>
+              <Stack spacing={400}>
+                <Text variant="body" size="medium">
+                  {initialContent}
+                </Text>
+              </Stack>
+            </ScrollContainer>
+          </ScrollContext>
+        </Popover>
+
+        <Button ref={triggerRef}>Open Popover</Button>
+      </div>
+    )
+  },
+}