+++ src/components/TabList/TabList.tsx
@@ -6,18 +6,17 @@ import './TabList.scss'
 /* --- */
 
 const TabListComponent = (
-  { id, className, variant = 'default', children, ...rest }: TabListProps,
+  { id, className, variant = 'default', scrollable = true, children, ...rest }: TabListProps,
   ref: preact.Ref<HTMLDivElement>,
 ) => {
   const _className = bem('TabList', undefined, {
     variant,
+    scrollable,
   })
 
   return (
     <div id={id} className={[_className, className].join(' ').trim()} ref={ref} {...rest}>
-      {children != null && children !== false && children !== true && (
-        <div className="TabList__children">{children}</div>
-      )}
+      {children != null && children !== false && children !== true && <div className="TabList__children">{children}</div>}
     </div>
   )
 }