+++ src/components/TabList/TabList.types.ts
@@ -2,5 +2,6 @@ export interface TabListProps {
   id?: string
   className?: string
   variant?: 'default' | 'list'
+  scrollable?: boolean
   children: preact.ComponentChildren
 }