+++ src/themes/figmaDark.ts
@@ -11,7 +11,7 @@ const figmaDark: ColorTokenTree = {
         'default-hover': '#FFFFFF0F',
         'default-pressed': '#FFFFFF1F',
         'default-selected': '#85A5FF5C',
-        'default-selected-hover': '#91B4FF66',
+        'default-selected-hover': '#91B4FF5C',
         'default-selected-pressed': '#85A5FF5C',
 
         secondary: '#383838',
@@ -169,7 +169,7 @@ const figmaDark: ColorTokenTree = {
         default: '#FFFFFF',
         secondary: '#A8D6F7',
 
-        disabled: '#FFFFFF',
+        disabled: '#2C2C2C',
       },
       border: {
         default: '#7CC4F8',